Update variable names and test archive code
There are some variables names that were confusing and not consistent
with the other naming conventions used. So this commit fixes those
variable names.

This commit also start the experimentation of the archive code for the
file to be uploaded. This is a work-in-progress, but this is a working
piece of code. This moves the archive code from a function and removes
the function for being too simple to be needed.